re-design popup page
