+++ src/popup/popup.css
@@ -3,8 +3,18 @@
 @tailwind utilities;
 
 body {
-    width: 250px;
-    height: 400px;
+    width: 400px;
+    height: 600px;
     padding: 15px;
-    background-color: white;
+    background-color: #0d1117;
+}
+
+.popup-tab {
+    @apply inline-block py-2 px-4 border-b-2 border-transparent cursor-pointer;
+}
+.popup-tab.active {
+    @apply text-blue-600 border-blue-600;
+}
+.popup-tab:hover {
+    @apply text-gray-300 border-gray-300;
 }
