--- src/services/gh-api.ts
@@ -1,5 +1,0 @@
-
-const api = () => {
-}
-
-export default api;
